requirements.txt
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.5
openai==1.12.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pydantic==2.5.3
fastapi==0.108.0
uvicorn==0.25.0
requests==2.31.0
python-dotenv==1.0.0
streamlit==1.29.0
pandas==2.1.4
numpy==1.24.3
tiktoken==0.5.2
chromadb==0.4.22
pytest==7.4.3
pytest-cov==4.1.0
python-multipart==0.0.6
aiofiles==23.2.1